minor fixes

+++ src/main/java/com/codecool/foodswap/repositories/FoodRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FoodRepository extends JpaRepository<Food, Long> {
+public interface FoodRepository extends JpaRepository<Food, Integer> {
 
     List<Food> findFoodsByGroup(Group group);
